# UOps.IF with rewrite rules