--- a/tinygrad/codegen/uopgraph.py
+++ b/tinygrad/codegen/uopgraph.py
@@ -427,23 +427,20 @@ expander = PatternMatcher([
 ])
 
 def create_gate(root:UOp) -> Optional[UOp]:
-  if len(root.src) == 3: return None
   @functools.lru_cache(None)
   def _gate_srcs(u:UOp, gate:UOp) -> UOp:
     if u.op is UOps.LOAD and u.src[-1].op is UOps.BARRIER: return UOp(u.op, u.dtype, u.src[:-1]+(UOp(UOps.IF, None, (gate, u.src[-1])),), u.arg)
-    if (replace_source:=tuple(_gate_srcs(x, gate) for x in u.src)) != u.src: return UOp(u.op, u.dtype, replace_source, u.arg)
-    return u
-  return ret if (ret:=_gate_srcs(root, root.src[3])) is not root else None
+    return u if (replace_source:=tuple(_gate_srcs(x, gate) for x in u.src)) == u.src else UOp(u.op, u.dtype, replace_source, u.arg)
+  return None if len(root.src) == 3 or (ret:=_gate_srcs(root, root.src[3])) is root else ret
 
 gate_creator = PatternMatcher([(NOp(UOps.STORE, name="root"), create_gate)])
 
 def delete_redundant_gates(root:UOp) -> Optional[UOp]:
-  if len(root.src) == 3: return None
   @functools.lru_cache(None)
   def find_gate(x:UOp) -> Optional[UOp]:
     if x.op is UOps.IF: return x
     return next((ret for s in x.src if (ret:=find_gate(s)) is not None), None)
-  if (gate:=find_gate(root)) is None or gate.src[0] is not root.src[3]: return None
+  if len(root.src) == 3 or (gate:=find_gate(root)) is None or gate.src[0] is not root.src[3]: return None
   return UOp(UOps.STORE, root.dtype, root.src[:3], root.arg)
 
 gate_folder = PatternMatcher([(NOp(UOps.STORE, name="root"), delete_redundant_gates)])