This is [more correct](https://github.com/tinygrad/tinygrad/actions/runs/10166804663/job/28117790767?pr=5812#step:24:94), no material performance diff:
master:
```
~/code/tinygrad> PROFILE=0 python3 test/external/external_benchmark_schedule.py
***** model forward in  58.11 ms
***** model schedule in  20.10 ms
***** model uops in 118.36 ms
***** model linearize in 1565.30 ms
```
this branch:
```
~/code/tinygrad> PROFILE=0 python3 test/external/external_benchmark_schedule.py
***** model forward in  58.09 ms
***** model schedule in  19.80 ms
***** model uops in 119.52 ms
***** model linearize in 1547.35 ms
```