@chenyuxyz 2024-07-30
saw this in metal diff
<img width="502" alt="image" src="https://github.com/user-attachments/assets/77c1cf71-1eaf-4626-8f33-db604ef664c7">

const True or false can be removed, you can use `vmax.arg == False` or `vmin.arg == True` to fold gate