@Qazalin 2024-07-30
Yea we have a UPat for that, it didn't do it before because I rewrote IF way too early. Now should be good.
https://github.com/tinygrad/tinygrad/blob/master/tinygrad/codegen/uopgraph.py#L299-L306