@chenyuxyz 2024-07-30
don't think that match though, the process replay diff are from this pr. i think that only matches if the whole gate clause is const

it's probably a good idea to fold `or False` and `and True` anyway even outside of load store