@Qazalin 2024-07-30
ahhh, the expander is creating ALUs outside of the graph that never get rewritten by `constant_folder` again.